requirements.txt
Django==2.1.2
gunicorn==19.9.0
whitenoise==4.1
djangorestframework==3.9.0